Correção dos exercícios da aula 19 e início da aula 20.

+++ src/cursojava/basico/aula19/exercicios/Exercicio36.java
@@ -9,10 +9,10 @@ public class Exercicio36 {
 //		ao expoente igual a posição do respectivo elemento, ou seja:
 //		A[i] = 2 i . Sugestão int A[11];
 		
-		double[] vetorA = new double[11];
+		int[] vetorA = new int[11];
 		
 		for(int i=0; i<vetorA.length; i++) {
-			vetorA[i] = Math.pow(2, i);
+			vetorA[i] = (int) Math.pow(2, i);
 			System.out.println(2 + " elevado a " + i + " = " + vetorA[i]);
 		}
 	}
